feat: no return of float

+++ DHT.h
@@ -64,13 +64,13 @@ class DHT {
 public:
   DHT(uint8_t pin, uint8_t type, uint8_t count = 6);
   void begin(uint8_t usec = 55);
-  float readTemperature(bool S = false, bool force = false);
+  int16_t readTemperature(bool S = false, bool force = false);
   float convertCtoF(float);
   float convertFtoC(float);
   float computeHeatIndex(bool isFahrenheit = true);
   float computeHeatIndex(float temperature, float percentHumidity,
                          bool isFahrenheit = true);
-  float readHumidity(bool force = false);
+  uint16_t readHumidity(bool force = false);
   bool read(bool force = false);
 
 private:
